pyproject.toml
[project]
name = "landry"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.1",
    "zeep>=4.3.1",
]
